Make templates require Scheme enum struct wrapping
- Pass `SchemeSystem` into Template creation to allow for different
  template builders to be built into tinted-builder-rust
- Remove anyhow crate from tinted-builder lib and use custom Error enum
  with thiserror library
