+++ tinted-builder/Cargo.toml
@@ -13,14 +13,16 @@ categories = ["command-line-utilities", "template-engine"]
 rust-version = "1.74.0"
 
 [dependencies]
-anyhow = "1.0.80"
 clap = "4.5.2"
 dirs = "5.0.1"
+quick-xml = { version = "0.36.1", features = ["serialize"] }
 regex = "1.10.4"
 ribboncurls = "0.2.1"
 serde = { version = "1.0.197", features = ["derive"] }
 serde_yaml = "0.9.32"
+thiserror = "1.0.63"
 unicode-normalization = "0.1.23"
 
 [dev-dependencies]
+anyhow = "1.0.86"
 strip-ansi-escapes = "0.2.0"
