--- tinted-builder/src/constants.rs
@@ -1,9 +1,0 @@
-pub(crate) const REQUIRED_BASE16_PALETTE_KEYS: [&str; 16] = [
-    "base00", "base01", "base02", "base03", "base04", "base05", "base06", "base07", "base08",
-    "base09", "base0A", "base0B", "base0C", "base0D", "base0E", "base0F",
-];
-pub(crate) const REQUIRED_BASE24_PALETTE_KEYS: [&str; 24] = [
-    "base00", "base01", "base02", "base03", "base04", "base05", "base06", "base07", "base08",
-    "base09", "base0A", "base0B", "base0C", "base0D", "base0E", "base0F", "base10", "base11",
-    "base12", "base13", "base14", "base15", "base16", "base17",
-];
